Updated intents
The intent scope is already marked with checks. I've tried training the program and it works flawlessly. I think we only need more patterns of questions for
it to be a viable chat bot like you said "Store the data in a database and retrain it". Also, the spacing is now done right lol. We do really need a bunch of
data.